Us 11114 : fixed declaration page declaration text placement (#181)
* adding declaration text to DefaultForm.tsx instead of checkbox.tsx

* fixed declaration page layout, added declaration texts to DefaultForm

* lint fixes
